+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed package resolution for frontend libraries (pbf, gtfs-realtime, maplibre-gl, anchorme) when using as a dependency by pre-copying files to `dist/frontend_libraries` during postinstall
+
 ## [2.12.1] - 2025-11-19
 
 ### Updated